The backend codes are cleaned up.
