+++ packages/backend/libs/Cargo.toml
@@ -7,6 +7,7 @@ edition = { workspace = true }
 icicle-runtime = { workspace = true }
 icicle-core = { workspace = true }
 icicle-bls12-381 = { workspace = true }
+icicle-hash = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 byteorder = { workspace = true }
