+++ packages/backend/prove/Cargo.toml
@@ -14,10 +14,6 @@ libs = { workspace = true }
 rayon = { workspace = true }
 bincode = { workspace = true }
 
-[[bin]]
-name = "prove"
-path = "src/main.rs"
-
 [features]
 default = []
 testing-mode = []
