+++ packages/backend/verify/verify-rust/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "verify"
+version = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+icicle-runtime = { workspace = true }
+icicle-core = { workspace = true }
+icicle-hash = { workspace = true }
+icicle-bls12-381 = { workspace = true }
+libs = { workspace = true }
+rayon = { workspace = true }
+bincode = { workspace = true }
+prove = { workspace = true }
+
+[features]
+default = []
+testing-mode = []
